# api/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: Marketplace service with products, orders, promo codes, and auth.

servers:
  - url: http://localhost:8080/api/v1

security:
  - BearerAuth: []

tags:
  - name: Auth
    description: Authentication and registration
  - name: Products
    description: Product CRUD operations
  - name: Orders
    description: Order management
  - name: PromoCodes
    description: Promo code management

paths:
  # ── Auth ──────────────────────────────────────────────
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Authenticate user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Products ──────────────────────────────────────────
  /products:
    post:
      tags: [Products]
      operationId: createProduct
      summary: Create a product (SELLER — own, ADMIN — any)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Products]
      operationId: listProducts
      summary: List products with pagination and filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-based)
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProductStatus"
          description: Filter by status
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category (exact match)
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

  /products/{id}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    get:
      tags: [Products]
      operationId: getProduct
      summary: Get product by ID
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Products]
      operationId: updateProduct
      summary: Update product (SELLER — own, ADMIN — any)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Products]
      operationId: deleteProduct
      summary: Soft-delete product — sets status to ARCHIVED (SELLER — own, ADMIN — any)
      responses:
        "204":
          description: Product archived
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Orders ────────────────────────────────────────────
  /orders:
    post:
      tags: [Orders]
      operationId: createOrder
      summary: Create an order (USER, ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreate"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Business rule violation (PRODUCT_INACTIVE, ORDER_HAS_ACTIVE, INSUFFICIENT_STOCK)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Promo code error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    get:
      tags: [Orders]
      operationId: getOrder
      summary: Get order by ID (USER — own, ADMIN — any)
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Orders]
      operationId: updateOrder
      summary: Update order items (USER — own, ADMIN — any). Only in CREATED state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Business rule violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Promo code error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    post:
      tags: [Orders]
      operationId: cancelOrder
      summary: Cancel an order (USER — own, ADMIN — any)
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/status:
    parameters:
      - $ref: "#/components/parameters/OrderId"
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      summary: Transition order status (follows state machine)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Promo Codes ───────────────────────────────────────
  /promo-codes:
    post:
      tags: [PromoCodes]
      operationId: createPromoCode
      summary: Create a promo code (SELLER, ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoCodeCreate"
      responses:
        "201":
          description: Promo code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCodeResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    OrderId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    # ── Auth ──────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 100
        role:
          $ref: "#/components/schemas/UserRole"

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user_id:
          type: integer
          format: int64
        role:
          $ref: "#/components/schemas/UserRole"

    UserRole:
      type: string
      enum: [USER, SELLER, ADMIN]

    # ── Product ───────────────────────────────────────
    ProductCreate:
      type: object
      required: [name, price, stock, category, status]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: double
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: "#/components/schemas/ProductStatus"

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: double
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: "#/components/schemas/ProductStatus"

    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: "#/components/schemas/ProductStatus"
        seller_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]

    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        totalElements:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # ── Order ─────────────────────────────────────────
    OrderCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/OrderItemCreate"
        promo_code:
          type: string
          pattern: "^[A-Z0-9_]{4,20}$"
          nullable: true

    OrderItemCreate:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
          maximum: 999

    OrderUpdate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/OrderItemCreate"

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/OrderStatus"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        promo_code:
          type: string
          nullable: true
        total_amount:
          type: number
          format: double
        discount_amount:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        product_id:
          type: integer
          format: int64
        quantity:
          type: integer
        price_at_order:
          type: number
          format: double

    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]

    # ── Promo Code ────────────────────────────────────
    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]{4,20}$"
        discount_type:
          $ref: "#/components/schemas/DiscountType"
        discount_value:
          type: number
          format: double
          minimum: 0.01
        min_order_amount:
          type: number
          format: double
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time

    PromoCodeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        discount_type:
          $ref: "#/components/schemas/DiscountType"
        discount_value:
          type: number
          format: double
        min_order_amount:
          type: number
          format: double
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean

    DiscountType:
      type: string
      enum: [PERCENTAGE, FIXED_AMOUNT]

    # ── Error ─────────────────────────────────────────
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"